Safari-609.1.11.1/LayoutTests/plugins/unavailable-plugin-indicator-obscurity-expected.txt

Expected to be obscured:
'tooSmall' was obscured: true (expected true; PASS)
'emptyRect' was obscured: true (expected true; PASS)
'lowOpacity' was obscured: true (expected true; PASS)
'lowOpacityParent' was obscured: true (expected true; PASS)
'stackedLowOpacityParents' was obscured: true (expected true; PASS)
'clippedByParent' was obscured: true (expected true; PASS)
'visibilityHidden' was obscured: true (expected true; PASS)
'visibilityHiddenOnParent' was obscured: true (expected true; PASS)
'pointerEventsBlocked' was obscured: true (expected true; PASS)
'pointerEventsBlockedOnParent' was obscured: true (expected true; PASS)
'pointerEventsBlockedOnObscuringSibling' was obscured: false (expected true; FAIL)
'leftHalfObscured' was obscured: true (expected true; PASS)
'rightHalfObscured' was obscured: true (expected true; PASS)
'topHalfObscured' was obscured: true (expected true; PASS)
'bottomHalfObscured' was obscured: true (expected true; PASS)
'centerObscured' was obscured: true (expected true; PASS)
'offScreenTopLeft' was obscured: true (expected true; PASS)
'offScreenTop' was obscured: true (expected true; PASS)
'offScreenLeft' was obscured: true (expected true; PASS)
'overflowScrollObscured' was obscured: true (expected true; PASS)
'smallIFrame' was obscured: true (expected true; PASS)

Expected to not be obscured:
'large' was obscured: false (expected false; PASS)
'stackedHighOpacityParents' was obscured: false (expected false; PASS)
'notClippedByParent' was obscured: false (expected false; PASS)
'topQuarterObscured' was obscured: false (expected false; PASS)
'bottomQuarterObscured' was obscured: false (expected false; PASS)
'largeIFrame' was obscured: false (expected false; PASS)

Passed all tests: false
